requirements.txt
Django
django-ckeditor
django-crispy-forms
geographiclib
geopy
gevent
Pillow
redis
requests
djangorestframework
django-adminlte-3
pyrebase